+++ qim3d/processing/__init__.py
@@ -1,10 +1,12 @@
 from ._layers import get_lines, segment_layers
 from ._local_thickness import local_thickness
 from ._structure_tensor import structure_tensor
+from ._fibers import FiberBundle
 
 __all__ = [
     'structure_tensor',
     'local_thickness',
     'get_lines',
     'segment_layers',
+    'FiberBundle'
 ]
